Separate directory branch widget into its own module.
