+++ kunquat/tracker/ui/views/dirbranch.py
@@ -1,0 +1,41 @@
+# -*- coding: utf-8 -*-
+
+#
+# Author: Tomi Jylhä-Ollila, Finland 2019
+#
+# This file is part of Kunquat.
+#
+# CC0 1.0 Universal, http://creativecommons.org/publicdomain/zero/1.0/
+#
+# To the extent possible under law, Kunquat Affirmers have waived all
+# copyright and related or neighboring rights to Kunquat.
+#
+
+from kunquat.tracker.ui.qt import *
+
+
+class DirBranch(QWidget):
+
+    def __init__(self, ui_model):
+        super().__init__()
+
+        self._ui_model = ui_model
+        self._current_dir = None
+
+        self._path = QLabel()
+
+        style_mgr = self._ui_model.get_style_manager()
+
+        h = QHBoxLayout()
+        margin = style_mgr.get_scaled_size_param('medium_padding')
+        h.setContentsMargins(margin, margin, margin, margin)
+        h.setSpacing(style_mgr.get_scaled_size_param('large_padding'))
+        h.addWidget(QLabel('Look in:'))
+        h.addWidget(self._path, 1)
+        self.setLayout(h)
+
+    def set_current_dir(self, new_dir):
+        self._current_dir = new_dir
+        self._path.setText(self._current_dir)
+
+
